pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "lailabot"
version = "0.1.0"
description = "Telegram-integrated Claude Code wrapper"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
lailabot = "lailabot.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
